==> code/1_setup.R <==
###################################################################
## Libraries
###################################################################
# library(rgeos) 
# library(rgdal)
# library(sp)
library(ggpubr)
library(tidyverse)
library(vegan)
library(tidybayes)
library(lme4)
library(lmerTest)
library(rstanarm)
library(rstantools)
library(glmmTMB)
